+++ src/lib/plsqlNavigator.ts
@@ -19,86 +19,99 @@ export class PlSqlNavigator {
                 rootSymbol: PLSQLSymbol,
                 rootSymbol2: PLSQLSymbol,
                 navigateSymbol: PLSQLSymbol,
+                navigateSymbol2: PLSQLSymbol,
                 isDeclaration: boolean,
                 packageName: string;
 
             // Declaration
             if (/*!cursorInfos.previousDot &&*/ this.isPackageDeclaration(cursorInfos.previousWord)) {
-              isDeclaration = true;
-              cursorSymbol = PlSqlParser.findSymbolByNameOffset(
-                  parserRoot.symbols, cursorInfos.currentWord, lineOffset);
+                isDeclaration = true;
+                cursorSymbol = PlSqlParser.findSymbolByNameOffset(
+                    parserRoot.symbols, cursorInfos.currentWord, lineOffset);
 
-              if (cursorSymbol && cursorSymbol.parent) {
-                  rootSymbol = PlSqlParser.switchSymbol(cursorSymbol.parent);
-                  if (rootSymbol && rootSymbol !== cursorSymbol.parent) {
-                      navigateSymbol = PlSqlParser.findSymbolByNameKind(rootSymbol.symbols, cursorSymbol.name, cursorSymbol.kind, false);
-                      return resolve(navigateSymbol);
-                  } else if (rootSymbol === cursorSymbol.parent)
-                      return resolve(); // No navigation here we are not in a package
-                  else
-                      // search in another file (spec && body are in separate files)
-                      packageName = cursorSymbol.parent.name;
-              } else
-                  // No parent => a function or a procedure not in a package
-                  return resolve();
+                if (cursorSymbol && cursorSymbol.parent) {
+
+                    // switch in body (spec and body are in body)
+                    if (cursorSymbol.parent.kind !== PLSQLSymbolKind.packageSpec) {
+                        navigateSymbol = PlSqlParser.findSymbolByNameKind(cursorSymbol.parent.symbols, cursorSymbol.name, PlSqlParser.switchSymbolKind(cursorSymbol.kind), false);
+                        if (navigateSymbol)
+                            return resolve(navigateSymbol);
+                    }
+
+                    // switch body <-> spec
+                    rootSymbol = PlSqlParser.switchSymbol(cursorSymbol.parent);
+                    if (rootSymbol && rootSymbol !== cursorSymbol.parent) {
+                        navigateSymbol = PlSqlParser.findSymbolByNameKind(rootSymbol.symbols, cursorSymbol.name, PlSqlParser.switchSymbolKind(cursorSymbol.kind), false);
+                        return resolve(navigateSymbol);
+                    } else if (rootSymbol === cursorSymbol.parent)
+                        return resolve(); // No navigation here we are not in a package
+                    else
+                        // search in another file (spec && body are in separate files)
+                        packageName = cursorSymbol.parent.name;
+                } else
+                    // No parent => a function or a procedure not in a package
+                    return resolve();
 
             // Call
             } else {
-              // Body => Body or Spec
-              rootSymbol = PlSqlParser.findSymbolNearOffset(parserRoot.symbols, lineOffset, false);
-              if (rootSymbol && rootSymbol.kind === PLSQLSymbolKind.packageSpec)
-                  return resolve(); // No navigation here we are in a spec
+                // Body => Body or Spec
+                rootSymbol = PlSqlParser.findSymbolNearOffset(parserRoot.symbols, lineOffset, false);
+                if (rootSymbol && rootSymbol.kind === PLSQLSymbolKind.packageSpec)
+                    return resolve(); // No navigation here we are in a spec
 
-              packageName = cursorInfos.previousDot ? cursorInfos.previousWord : '';
-              // Use synonyme for package
-              if (pkgGetName_cb)
-                  packageName = pkgGetName_cb.call(this, packageName);
+                packageName = cursorInfos.previousDot ? cursorInfos.previousWord : '';
+                // Use synonyme for package
+                if (pkgGetName_cb)
+                    packageName = pkgGetName_cb.call(this, packageName);
 
-              // Search in current file
-              if (rootSymbol && (!packageName || (packageName.toLowerCase() === rootSymbol.name.toLowerCase()))) {
-                  // Search in current body of file
-                  navigateSymbol = PlSqlParser.findSymbolByNameOffset(rootSymbol.symbols, cursorInfos.currentWord, 0, false);
-                  if (navigateSymbol)
-                      return resolve(navigateSymbol);
-                  // Search in current spec (maybe a constant or type definition)
-                  rootSymbol2 = PlSqlParser.switchSymbol(rootSymbol);
-                  if (rootSymbol2 && rootSymbol2 !== rootSymbol) {
-                      navigateSymbol = PlSqlParser.findSymbolByNameOffset(rootSymbol2.symbols, cursorInfos.currentWord, 0, false);
-                      if (navigateSymbol)
-                          return resolve(navigateSymbol);
-                  } else if (!packageName && !rootSymbol2 && rootSymbol.kind === PLSQLSymbolKind.packageBody) {
-                      // spec is in separeate file
-                      packageName = rootSymbol.name;
-                  }
-              }
+                // Search in current file
+                if (rootSymbol && (!packageName || (packageName.toLowerCase() === rootSymbol.name.toLowerCase()))) {
+                    // Search in current body of file  (recursive for subFunctions or subProcedure)
+                    navigateSymbol = PlSqlParser.findSymbolByNameOffset(rootSymbol.symbols, cursorInfos.currentWord, 0, true);
+                    if (navigateSymbol) {
+                        if (PlSqlParser.isSymbolSpec(navigateSymbol))
+                            navigateSymbol2 = PlSqlParser.findSymbolByNameKind(rootSymbol.symbols, navigateSymbol.name, PlSqlParser.switchSymbolKind(navigateSymbol.kind), false);
+                        return resolve(navigateSymbol2 || navigateSymbol);
+                    }
+                    // Search in current spec (maybe a constant or type definition)
+                    rootSymbol2 = PlSqlParser.switchSymbol(rootSymbol);
+                    if (rootSymbol2 && rootSymbol2 !== rootSymbol) {
+                        navigateSymbol = PlSqlParser.findSymbolByNameOffset(rootSymbol2.symbols, cursorInfos.currentWord, 0, false);
+                        if (navigateSymbol)
+                           return resolve(navigateSymbol);
+                    } else if (!packageName && !rootSymbol2 && rootSymbol.kind === PLSQLSymbolKind.packageBody) {
+                        // spec is in separate file
+                        packageName = rootSymbol.name;
+                    }
+                }
             }
 
             // Search in external files
             const search = {
-              package: packageName,
-              cursorWord: cursorInfos.currentWord,
-              isDeclaration: isDeclaration
+                package: packageName,
+                cursorWord: cursorInfos.currentWord,
+                isDeclaration: isDeclaration
             };
             let files;
             this.getGlobFiles(this.getGlobCmd(search, search_cb))
-              .then(globFiles => {
-                  files = globFiles;
-                  return this.parseFiles(files, search);
-              })
-             .then(symbol => {
-                  // search without packageName (because it's perhaps only the name of the schema)
-                  if (!symbol && search.package) {
-                      search.package = null;
-                      return this.parseFiles(files, search);
-                  }
-                  return symbol;
-              })
-              .then(symbol => {
-                  resolve(symbol);
-              })
-              .catch(err => {
-                  reject(err);
-              });
+                .then(globFiles => {
+                    files = globFiles;
+                    return this.parseFiles(files, search, this.gotoFile);
+                })
+               .then(symbol => {
+                    // search without packageName (because it's perhaps only the name of the schema)
+                    if (!symbol && search.package) {
+                        search.package = null;
+                        return this.parseFiles(files, search, this.gotoFile);
+                    }
+                    return symbol;
+                })
+                .then(symbol => {
+                    resolve(symbol);
+                })
+                .catch(err => {
+                    reject(err);
+                });
         });
     }
 
@@ -130,6 +143,29 @@ export class PlSqlNavigator {
         };
     }
 
+    public static complete(cursorInfos: PLSQLCursorInfos, pkgGetName_cb, search_cb): Promise<PLSQLSymbol[]> {
+
+        return new Promise<PLSQLSymbol[]>((resolve, reject) => {
+
+            const search = {
+                package: cursorInfos.previousWord,
+                cursorWord: cursorInfos.currentWord
+            };
+            let files;
+            this.getGlobFiles(this.getGlobCmd(search, search_cb))
+                .then(globFiles => {
+                    files = globFiles;
+                    return this.parseFiles(files, search, this.completeItem);
+                })
+                .then(symbols => {
+                    return resolve(symbols);
+                })
+                .catch(err => {
+                    reject(err);
+                });
+        });
+    }
+
     private static isPackageDeclaration(text) {
         return text && ['function', 'procedure'].includes(text.toLowerCase());
     }
@@ -152,7 +188,8 @@ export class PlSqlNavigator {
         let files: string[] = [];
         if (searchTexts.package)
             files.push(searchTexts.package);
-        files.push(searchTexts.cursorWord);
+        if (searchTexts.cursorWord)
+            files.push(searchTexts.cursorWord);
 
         let search = {
             files: files,
@@ -175,7 +212,7 @@ export class PlSqlNavigator {
         return search;
     }
 
-    private static parseFiles(files: string[], searchInfos): Promise<PLSQLSymbol> {
+    private static parseFiles(files: string[], searchInfos, func): Promise<any> {
 
         return new Promise((resolve, reject) => {
 
@@ -188,7 +225,7 @@ export class PlSqlNavigator {
 
                 me.readFile(files[index])
                     .then(rootSymbol => {
-                        const navigateSymbol = me.gotoFile(searchInfos, rootSymbol);
+                        const navigateSymbol = func.call(this, searchInfos, rootSymbol);
 
                         if (navigateSymbol === null)
                             process(index + 1);
@@ -203,6 +240,16 @@ export class PlSqlNavigator {
             })(0);
 
         });
+    }
+
+    private static completeItem(searchInfos, rootSymbol: PLSQLRoot): PLSQLSymbol[] {
+        let symbols;
+        const parentSymbol = PlSqlParser.findSymbolByNameKind(rootSymbol.symbols, searchInfos.package, [PLSQLSymbolKind.packageSpec], false);
+        if (parentSymbol)
+            symbols = parentSymbol.symbols;
+        else
+            symbols = null; // return null for continue search with next file
+        return symbols;
     }
 
     private static gotoFile(searchInfos, rootSymbol: PLSQLRoot): PLSQLSymbol {
@@ -222,7 +269,7 @@ export class PlSqlNavigator {
             if (navigateSymbol) {
                 // if function/procedure in spec => search in body
                 if (!searchInfos.isDeclaration && navigateSymbol.parent && navigateSymbol.parent.kind === PLSQLSymbolKind.packageSpec &&
-                    [PLSQLSymbolKind.function, PLSQLSymbolKind.procedure].includes(navigateSymbol.kind)) {
+                    [PLSQLSymbolKind.functionSpec, PLSQLSymbolKind.procedureSpec].includes(navigateSymbol.kind)) {
                     parentSymbol = PlSqlParser.switchSymbol(navigateSymbol.parent);
                     if (parentSymbol !== navigateSymbol.parent)
                         return PlSqlParser.findSymbolByNameOffset(parentSymbol.symbols, searchInfos.cursorWord, 0, false);