+++ src/plsqlDefinition.provider.ts
@@ -1,10 +1,8 @@
 import * as vscode from 'vscode';
-import * as path from 'path';
 
-import PLSQLSettings from './plsql.settings';
 import PlSqlParser from './plsqlParser.vscode';
-import { PlSqlNavigator } from './lib/plsqlNavigator';
-import { PLSQLCursorInfos } from './lib/plsqlNavigator';
+
+import { PlSqlNavigatorVSC as PlSqlNavigator }  from './plsqlNavigator.vscode';
 
 export class PLSQLDefinitionProvider implements vscode.DefinitionProvider {
 
@@ -12,14 +10,7 @@ export class PLSQLDefinitionProvider implements vscode.DefinitionProvider {
 
         return new Promise<vscode.Location>((resolve, reject) => {
 
-            PlSqlParser.initParser(PLSQLSettings.getCommentInSymbols());
-
-            const line = document.lineAt(position),
-                  cursorInfos = this.getCursorInfos(document, line, position),
-                  parserRoot = PlSqlParser.parseDocument(document);
-
-            PlSqlNavigator.goto(cursorInfos, document.offsetAt(line.range.start), parserRoot,
-                    this.translatePackageName.bind(this, document), this.getGlobCmdEx.bind(this, document))
+            PlSqlNavigator.goto(document, position)
                 .then(symbol => {
                     return this.getFileLocation(symbol);
                 })
@@ -31,35 +22,6 @@ export class PLSQLDefinitionProvider implements vscode.DefinitionProvider {
                 });
         });
     };
-
-    private getCursorInfos(document: vscode.TextDocument, line, position: vscode.Position): PLSQLCursorInfos {
-
-        const lineText = line.text,
-              range = document.getWordRangeAtPosition(position),
-              currentWord = document.getText(range);
-
-        return PlSqlNavigator.getCursorInfos(currentWord, range.end.character, lineText);
-    }
-
-    private translatePackageName(document, packageName: string): string {
-        return PLSQLSettings.translatePackageName(packageName);
-    }
-
-    private getGlobCmdEx(document, search) {
-        const {cwd, ignore} = PLSQLSettings.getSearchInfos(document.uri);
-        // Ignore current file
-        ignore.push(path.relative(cwd, document.uri.fsPath));
-
-        return {
-            files: search.files,
-            ext: PLSQLSettings.getSearchExt(search.ext),
-            params: {
-                nocase: true,
-                cwd: cwd,
-                ignore: ignore
-            }
-        };
-    }
 
     private getFileLocation(navigateSymbol: PLSQLSymbol): Promise<vscode.Location> {
         return new Promise((resolve, reject) => {